requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
